added icons

+++ stack.py
@@ -1,16 +1,22 @@
 import yaml
-import os, time
+import os
+import time
 import logging
-import subprocess, threading, psutil, signal
+import subprocess
+import threading
+import psutil
+import signal
 from plugins import flask, npm
+from plugins._base import BasePlugin
 
 
 def process_variables(s: str, vars: dict) -> str:
     # TODO: proper regex?
     for k, v in vars.items():
         s = s.replace('$'+k, v)
-    
+
     return s
+
 
 class ServerConfig:
 
@@ -19,11 +25,12 @@ class ServerConfig:
     def __init__(self, data):
         self.data = data
 
-    def get(self, k, d = None)->any:
+    def get(self, k, d=None) -> any:
         if k in self.data:
             return self.data[k]
         else:
             return d
+
 
 class Config:
 
@@ -31,10 +38,11 @@ class Config:
 
     def __init__(self):
         self.servers = {}
-    
+
     @staticmethod
     def get_config_file():
-        f = os.path.realpath(os.path.join(os.path.dirname(__file__), 'config.yaml'))
+        f = os.path.realpath(os.path.join(
+            os.path.dirname(__file__), 'config.yaml'))
 
         if os.path.exists(f):
             return f
@@ -42,7 +50,7 @@ class Config:
             raise IOError('Config file "%s" not found' % f)
 
     def load_yaml(self, filename: str):
-        
+
         with open(filename, 'r') as fp:
             try:
                 servers = yaml.load(fp, Loader=yaml.SafeLoader)['servers']
@@ -51,11 +59,12 @@ class Config:
                     self.servers[name] = ServerConfig(conf)
 
             except Exception as e:
-                raise Exception('Could not load config file "%s": %s' % (filename, str(e)))
-
+                raise Exception(
+                    'Could not load config file "%s": %s' % (filename, str(e)))
 
     def get_servers(self) -> dict:
         return self.servers
+
 
 class ServerRunnerEnvironment:
 
@@ -74,19 +83,22 @@ class ServerRunner(threading.Thread):
     process_id = None
     name = ''
     logger = None
+    shutdown_strategy = None
 
-    def __init__(self, conf: ServerConfig, env: ServerRunnerEnvironment, name = 'Unnamed'):
+    def __init__(self, conf: ServerConfig, env: ServerRunnerEnvironment, name='Unnamed'):
         super().__init__()
 
         self.conf = conf
         self.env = env
         self.name = name
+        self.output = []
 
         # Setup logging
         self.logger = logging.getLogger(name)
         self.logger.setLevel(logging.DEBUG)
         hdl = logging.StreamHandler()
-        hdl.setFormatter(logging.Formatter(fmt='%(levelname)s:%(name)s:%(message)s'))
+        hdl.setFormatter(logging.Formatter(
+            fmt='%(levelname)s:%(name)s:%(message)s'))
         self.logger.addHandler(logging.StreamHandler())
 
     def run(self):
@@ -97,16 +109,19 @@ class ServerRunner(threading.Thread):
         # find plugin to handle the config
         for plugin in self.env.plugins:
             if self.conf.get('type', '') in plugin.handles:
-                self.logger.info('Using "%s" for type "%s"' % (plugin, self.conf.get('type')))
+                self.logger.info('Using "%s" for type "%s"' %
+                                 (plugin, self.conf.get('type')))
                 plugin_inst = plugin(self.conf)
                 cmd = plugin_inst.get_command()
+                self.shutdown_strategy = BasePlugin.SHUTDOWN_STRATEGY_INT
 
                 # stop looking for plugin
                 found_plugin = True
                 break
 
         if not found_plugin:
-            raise Exception('Could not find a plugin to handle type "%s"' % self.conf.get('type'))
+            raise Exception(
+                'Could not find a plugin to handle type "%s"' % self.conf.get('type'))
 
         cwd = os.path.realpath(self.conf.get('dir', '.'))
 
@@ -115,9 +130,8 @@ class ServerRunner(threading.Thread):
             venv_dir = os.path.join(venv_dir, 'bin', 'activate')
             cmd = 'source %s && %s' % (venv_dir, cmd)
 
-        
         self.logger.debug('Running command "%s"' % cmd)
-       
+
         proc = subprocess.Popen(
             args=cmd,               # Python doc recommends str arg if shell=True
             shell=True,
@@ -134,13 +148,14 @@ class ServerRunner(threading.Thread):
             if (line := proc.stdout.readline()) != '':
                 self.output.append(line)
 
-        self.logger.info('Server process stopped.')
+        self.logger.info('Server process ended with "%s".' % res)
 
 
 # we need a separate function to shutdown the process, since the thread will block on stdout.readline()
 def stop_runner(runner: ServerRunner):
 
-    if not runner.is_alive(): return
+    if not runner.is_alive():
+        return
 
     try:
         runner.logger.info('Shutting down process...')
@@ -149,12 +164,17 @@ def stop_runner(runner: ServerRunner):
         children = parent.children(recursive=True)
 
         for child in children:
-            child.send_signal(signal.SIGINT)
+            if runner.shutdown_strategy == BasePlugin.SHUTDOWN_STRATEGY_INT:
+                child.send_signal(signal.SIGINT)
+            else:
+                runner.logger.error(
+                    'Unknown shutdown strategy. Killing processes.')
 
         gone, alive = psutil.wait_procs(children, timeout=3)
 
         for still_alive in alive:
-            runner.logger.warning('Process %s is still alive, sending KILL signal...' % still_alive)
+            runner.logger.warning(
+                'Process %s is still alive, sending KILL signal...' % still_alive)
             sill_alive.kill()
 
     except psutil.NoSuchProcess:
